+++ finagle-examples/src/main/scala/ru/tinkoff/tschema/example/package.scala
@@ -1,0 +1,9 @@
+package ru.tinkoff.tschema
+import ru.tinkoff.tschema.finagle._
+import ru.tinkoff.tschema.finagle.zioInstance._
+import scalaz.zio._
+
+package object example {
+  type Http[+A]    = ZIOHttp[ExampleEnv, Nothing, A]
+  type Example[+A] = ZIO[ExampleEnv, Nothing, A]
+}