+++ modules/finagle-zio/src/main/scala/ru/tinkoff/tschema/finagle/zioInstance.scala
@@ -3,7 +3,7 @@ import cats.Monad
 import cats.data.NonEmptyMap
 import cats.instances.string._
 import cats.syntax.semigroup._
-import com.twitter.finagle.http.{Request, Response}
+import com.twitter.finagle.http.{Request, Response, Status}
 import com.twitter.finagle.{Service, http}
 import com.twitter.util.{Future, Promise}
 import ru.tinkoff.tschema.utils.SubString
@@ -20,59 +20,62 @@ object zioInstance {
   final case class ZioRouting[+R](
       request: http.Request,
       path: CharSequence,
-      matched: Ref[Int],
+      matched: Int,
       embedded: R
   )
 
   type ZIOHttp[-R, +E, +A] = ZIO[ZioRouting[R], Fail[E], A]
   type ZIORoute[+A]        = ZIOHttp[Any, Nothing, A]
 
-  implicit val zioRouted: Routed[ZIORoute] =
-    new Routed[ZIORoute] {
-      val FMonad: Monad[ZIORoute] = zio.interop.catz.ioInstances
+  def zioRouted[R, E]: RoutedPlus[ZIOHttp[R, E, ?]] =
+    new RoutedPlus[ZIOHttp[R, E, ?]] {
+      val FMonad: Monad[ZIOHttp[R, E, ?]] = zio.interop.catz.ioInstances
 
-      def matched: ZIORoute[Int]                = ZIO.accessM(_.matched.get)
-      def setMatched[A](m: Int): ZIORoute[Unit] = ZIO.accessM(_.matched.set(m))
-      def path: ZIORoute[CharSequence]          = ZIO.access(_.path)
-      def request: ZIORoute[http.Request]       = ZIO.access(_.request)
-      def reject[A](rejection: Rejection): ZIORoute[A] =
+      def matched: ZIORoute[Int] = ZIO.access(_.matched)
+
+      def withMatched[A](m: Int, fa: ZIOHttp[R, E, A]): ZIOHttp[R, E, A] = fa.provideSome(_.copy(matched = m))
+
+      override def addMatched[A](i: Int, fa: ZIOHttp[R, E, A]): ZIOHttp[R, E, A] =
+        fa.provideSome(r => r.copy(matched = r.matched + i))
+      def path: ZIORoute[CharSequence]    = ZIO.access(_.path)
+      def request: ZIORoute[http.Request] = ZIO.access(_.request)
+      def reject[A](rejection: Rejection): ZIOHttp[R, E, A] =
         unmatchedPath.flatMap(path => throwRej(rejection withPath path.toString))
 
-      def combineK[A](x: ZIORoute[A], y: ZIORoute[A]): ZIORoute[A] =
-        matched >>= (m => catchRej(x)(xrs => setMatched(m) *> catchRej(y)(yrs => throwRej(xrs |+| yrs))))
+      def combineK[A](x: ZIOHttp[R, E, A], y: ZIOHttp[R, E, A]): ZIOHttp[R, E, A] =
+        catchRej(x)(xrs => catchRej(y)(yrs => throwRej(xrs |+| yrs)))
     }
 
   def zioRunnable[R, E <: Throwable](
       rejectionHandler: Rejection.Handler = Rejection.defaultHandler): Runnable[ZIOHttp[R, E, ?], ZIO[R, E, ?]] =
     new Runnable[ZIOHttp[R, E, ?], ZIO[R, E, ?]] {
       def run(zioResponse: ZIOHttp[R, E, Response]): ZIO[R, E, Service[Request, Response]] =
-        for {
-          ref     <- Ref.make(0)
-          runtime <- ZIO.runtime[R]
-        } yield execResponse(ref, runtime, rejectionHandler, zioResponse, _)
+        ZIO.runtime[R].flatMap(runtime => ZIO.effectTotal(execResponse(runtime, rejectionHandler, zioResponse, _)))
 
       def apply[A](fa: ZIO[R, E, A]): ZIOHttp[R, E, A] = fa.mapError(OtherFail(_)).provideSome(_.embedded)
     }
 
-  @inline private[this] def catchRej[A](z: ZIORoute[A])(f: Rejection => ZIORoute[A]): ZIORoute[A] =
+  @inline private[this] def catchRej[R, E, A](z: ZIOHttp[R, E, A])(f: Rejection => ZIOHttp[R, E, A]): ZIOHttp[R, E, A] =
     z.catchSome { case Rejected(xrs) => f(xrs) }
 
-  @inline private[this] def throwRej[A](map: Rejection): ZIORoute[A] = ZIO.fail(Rejected(map))
+  @inline private[this] def throwRej[R, E, A](map: Rejection): ZIOHttp[R, E, A] = ZIO.fail(Rejected(map))
 
-  @inline private[this] def execResponse[R, E <: Throwable](matchRef: Ref[Int],
-                                                            runtime: zio.Runtime[R],
+  @inline private[this] def execResponse[R, E <: Throwable](runtime: zio.Runtime[R],
                                                             handler: Rejection.Handler,
                                                             zioResponse: ZIOHttp[R, E, Response],
                                                             request: Request): Future[Response] = {
     val promise = Promise[Response]
     runtime.unsafeRunAsync(
-      zioResponse.provideSome[R](r => ZioRouting(request, SubString(request.path), matchRef, r)).catchAll {
+      zioResponse.provideSome[R](r => ZioRouting(request, SubString(request.path), 0, r)).catchAll {
         case Rejected(rejection) => ZIO.succeed(handler(rejection))
         case OtherFail(e)        => ZIO.fail(e)
       }
     ) {
-      case Exit.Success(resp)  => promise.setValue(resp)
-      case Exit.Failure(cause) => promise.raise(cause.squash)
+      case Exit.Success(resp) => promise.setValue(resp)
+      case Exit.Failure(cause) =>
+        val resp = Response(Status.InternalServerError)
+        resp.setContentString(cause.squash.getMessage)
+        promise.setValue(resp)
     }
     promise
   }