+++ modules/finagle/src/main/scala/ru/tinkoff/tschema/finagle/MkService.scala
@@ -1,49 +1,48 @@
 package ru.tinkoff.tschema.finagle
-import cats.{FlatMap, Monad}
-import com.twitter.finagle.http
+import cats.syntax.semigroupk._
+import cats.{FlatMap, SemigroupK}
 import com.twitter.finagle.http.Response
 import ru.tinkoff.tschema.macros.MakerMacro
-import shapeless.HList
-import cats.syntax.semigroupk._
-import cats.syntax.flatMap._
-import cats.syntax.apply._
 import ru.tinkoff.tschema.typeDSL.DSLDef
+import shapeless.HList
 
 object MkService {
-  def apply[F[_], Def <: DSLDef, Impl](definition: => Def)(impl: Impl): F[Response] =
-    macro MakerMacro.makeRouteHNil[macroInterface.type, Def, Impl, F[Response]]
+  def apply[F[_]] = new MkApply[F]
 
-  def of[F[_], Def <: DSLDef, Impl, In <: HList](definition: => Def)(impl: Impl)(input: In): F[Response] =
-    macro MakerMacro.makeRoute[macroInterface.type, Def, Impl, F[Response], In]
+  class MkApply[F[_]] {
+    def apply[Def <: DSLDef, Impl](definition: => Def)(impl: Impl): F[Response] =
+      macro MakerMacro.makeRouteHNil[F, macroInterface.type, Def, Impl, F[Response]]
+
+    def of[Def <: DSLDef, Impl, In <: HList](definition: => Def)(impl: Impl)(input: In): F[Response] =
+      macro MakerMacro.makeRoute[F, macroInterface.type, Def, Impl, F[Response], In]
+  }
 
   object macroInterface {
-    import Routed.implicits._
+    def makeResult[F[_], Out]: ResultPA1[F, Out] = new ResultPA1[F, Out]
 
-    def makeResult[Out]: ResultPA1[Out] = new ResultPA1[Out]
-
-    def concatResults[F[_]: Routed](x: F[Response], y: F[Response]): F[Response] =
+    def concatResults[F[_]: SemigroupK](x: F[Response], y: F[Response]): F[Response] =
       x combineK y
 
-    def serve[T] = new ServePA[T]
+    def serve[F[_], T] = new ServePA[F, T]
 
-    def route[Res](res: => Res) = new RoutePA(res)
+    def route[Res](res: => Res) = new RoutePA[Res](res)
 
-    class ResultPA1[Out] {
-      def apply[F[_], In <: HList, Impl](in: In)(impl: Impl)(key: String): F[Response] =
-        macro MakerMacro.makeResult[In, Out, Impl, F[Response]]
+    class ResultPA1[F[_], Out] {
+      def apply[In <: HList, Impl](in: In)(impl: Impl)(key: String): F[Response] =
+        macro MakerMacro.makeResult[F, In, Out, Impl, F[Response]]
     }
 
     class RoutePA[Res](res: => Res) {
       def apply[F[_]: Routed, In, Out](in: In)(implicit complete: Complete[F, Res]): F[Response] =
-        Routed.checkPathEnd *> complete.complete(res)
+        Routed.checkPathEnd(complete.complete(res))
     }
 
-    class ServePA[T] {
-      def apply[F[_]: FlatMap, In, Out](in: In)(implicit serve: Serve[T, F, In, Out]) = new ServePA2[F, T, In, Out](in)(serve)
+    class ServePA[F[_], T] {
+      def apply[In, Out](in: In)(implicit serve: Serve[T, F, In, Out], F: FlatMap[F]) = new ServePA2[F, T, In, Out](in)(serve)
     }
 
     class ServePA2[F[_]: FlatMap, T, In, Out](val in: In)(srv: Serve[T, F, In, Out]) {
-      def apply(f: Out => F[Response]): F[Response] = srv.process(in).flatMap(f)
+      def apply(f: Out => F[Response]): F[Response] = srv.process(in, f)
     }
   }
 }